+++ packages/gatsby/cache-dir/production-app.js
@@ -120,9 +120,7 @@ apiRunnerAsync(`onClientEntry`).then(() => {
   publicLoader.loadPage(browserLoc.pathname).then(page => {
     if (!page || page.status === `error`) {
       throw new Error(
-        `page resources for ${
-          browserLoc.pathname
-        } not found. Not rendering React`
+        `page resources for ${browserLoc.pathname} not found. Not rendering React`
       )
     }
     const Root = () => (