+++ packages/gatsby-source-shopify/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.9...gatsby-source-shopify@3.0.10) (2019-09-01)
+
+### Bug Fixes
+
+- update minor updates in packages except react, babel and eslint ([#17254](https://github.com/gatsbyjs/gatsby/issues/17254)) ([252d867](https://github.com/gatsbyjs/gatsby/commit/252d867))
+
+## [3.0.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.8...gatsby-source-shopify@3.0.9) (2019-08-24)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
+## [3.0.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.7...gatsby-source-shopify@3.0.8) (2019-08-24)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
+## [3.0.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.6...gatsby-source-shopify@3.0.7) (2019-08-23)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
+## [3.0.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.5...gatsby-source-shopify@3.0.6) (2019-08-23)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
+## [3.0.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.4...gatsby-source-shopify@3.0.5) (2019-08-22)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
+## [3.0.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.3...gatsby-source-shopify@3.0.4) (2019-08-22)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
+## [3.0.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.2...gatsby-source-shopify@3.0.3) (2019-08-21)
+
+### Bug Fixes
+
+- update dependency chalk to ^2.4.2 ([#16830](https://github.com/gatsbyjs/gatsby/issues/16830)) ([3c89607](https://github.com/gatsbyjs/gatsby/commit/3c89607))
+- update dependency graphql-request to ^1.8.2 ([#16886](https://github.com/gatsbyjs/gatsby/issues/16886)) ([1e6c29a](https://github.com/gatsbyjs/gatsby/commit/1e6c29a))
+
 ## [3.0.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.1...gatsby-source-shopify@3.0.2) (2019-08-21)
 
 **Note:** Version bump only for package gatsby-source-shopify